pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "py-cord>=2.7.0",
    "openai>=2.17.0",
    "python-dotenv>=1.2.1",
    "ollama>=0.6.1",
    "langchain-community",
    "langchain-text-splitters",
    "huggingface_hub",
    "faiss-cpu",
    "unstructured",
    "unstructured[docx,pdf]",
    "pypdf",
    "sentence-transformers"
]
